requirements.txt
python-dotenv==1.0.0
google-generativeai==0.3.1
streamlit==1.28.0
moviepy==1.0.3
opencv-python==4.8.1.78
numpy==1.26.0
pillow==10.1.0
pytubefix==8.12.2
requests==2.31.0
watchdog==3.0.0 